Admin - Move int tests to separate file
